Move some parts to domains
