+++ packages/domains/src/tx/operations.ts
@@ -15,40 +15,10 @@
 
 import { Class, Doc, DocumentValue, Model, Ref, Tx } from '@anticrm/core'
 import {
-  CORE_CLASS_OBJECTTX_DETAILS, CORE_CLASS_SPACE, CORE_CLASS_TX_OPERATION, CORE_CLASS_VDOC, CORE_MIXIN_SHORTID, ObjectTx, ObjectTxDetails, Space,
+  CORE_CLASS_OBJECTTX_DETAILS, CORE_CLASS_SPACE, CORE_CLASS_TX_OPERATION, CORE_CLASS_VDOC, CORE_MIXIN_SHORTID, ObjectTx, ObjectTxDetails, OperationProtocol, Space,
   txBuilder, TxBuilder, TxOperation, TxOperationKind, UpdateTx
 } from '@anticrm/domains'
 import { newCreateTx, newDeleteTx, newUpdateTx } from './tx'
-
-/**
- * Define operations with object modifications.
- */
-export interface OperationProtocol {
-  /**
-   * Perform creation of new document and store it into storage.
-   * Object ID will be automatically generated and assigned to object.
-   */
-  create: <T extends Doc>(_class: Ref<Class<T>>, values: DocumentValue<T>) => Promise<T>
-
-  /**
-   * Perform update of document properties.
-   */
-  update: <T extends Doc>(doc: T, value: Partial<Omit<T, keyof Doc>>) => Promise<T>
-
-  /**
-   * Perform update of document/embedded document properties using a builder pattern.
-   *
-   * It is possible to do a set, pull, push for different field values.
-   *
-   * push and pull are applicable only for array attributes.
-   */
-  updateWith: <T extends Doc>(doc: T, builder: (s: TxBuilder<T>) => TxOperation | TxOperation[]) => Promise<T>
-
-  /**
-   * Perform remove of object.
-   */
-  remove: <T extends Doc>(doc: T) => Promise<T>
-}
 
 function getSpace (model: Model, doc: Doc): { _objectSpace: Ref<Space> | undefined, spaceIsRequired: boolean } {
   if (model.is(doc._class, CORE_CLASS_SPACE)) {
