+++ plugins/recruiting/src/components/CandidateList.svelte
@@ -15,7 +15,7 @@ limitations under the License.
 <script lang="ts">
   import { createEventDispatcher } from 'svelte'
 
-  import { QueryUpdater } from '@anticrm/platform-core'
+  import type { QueryUpdater } from '@anticrm/presentation'
   import type { Space } from '@anticrm/domains'
   import { liveQuery } from '@anticrm/presentation'
