+++ server/server/src/workspace.ts
@@ -25,7 +25,7 @@ import { ReferenceIndex } from '@anticrm/domains/src/indices/reference'
 import { TitleIndex } from '@anticrm/domains/src/indices/title'
 import { TxIndex } from '@anticrm/domains/src/indices/tx'
 import { VDocIndex } from '@anticrm/domains/src/indices/vdoc'
-import { ClientTxStorage } from '@anticrm/client/src/clienttx'
+import { ClientTxStorage } from '@anticrm/domains/src/tx/clienttx'
 import { Collection, MongoClient, SortOptionObject, UpdateOneOptions, UpdateQuery } from 'mongodb'
 import type { FindAndModifyWriteOpResultObject } from 'mongodb'
 import { createPullArrayFilters, createPushArrayFilters, createSetArrayFilters } from './mongo_utils'
